Black mode add

+++ src/pages/Dashboard/index.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react"
 import moment from 'moment/min/moment-with-locales'
 import Loading from "../../components/Loading"
 import api from "../../services/api"
+import "./styles.css"
 moment.locale("pt-br")
 
 export default function Dashboard({ history }){
@@ -19,6 +20,7 @@ export default function Dashboard({ history }){
         }
         load()
     }, [history])
+    
     async function validate(event){
         event.preventDefault()
         if(saldo === "" || isNaN(parseInt(saldo.replace(",", ".")))){
@@ -35,26 +37,31 @@ export default function Dashboard({ history }){
         setUserSaldo(res.data.saldo)
     }
     async function resetar(){
+        if(!window.confirm("Você tem certeza que deseja resetar seu saldo?")){
+            return
+        }
         var res = await api.delete("/dashboard", { headers: { _id: localStorage.getItem("user") }})
         if(res.data.error){
             return alert(res.data.error)
         }
         setUserSaldo(res.data.saldo)
     }
+    async function logout(){
+        localStorage.removeItem("user")
+        history.push("/")
+    }
     return(
         <>
         {(loaded === true) ?
 <> 
-        <button onClick={() => {
-        localStorage.removeItem("user")
-        history.push("/")
-        }}>Logout</button>
+        <button onClick={() => logout()}>Logout</button>
         <form onSubmit={validate}>
                 <h1 id="h2">Seu saldo:</h1>
-                <h2 id="saldo">{parseFloat(userSaldo).toFixed(2).replace(".", ",")}</h2>
-                <h3>Quanto?</h3>
+                <h2 id="saldo">R${parseFloat(userSaldo).toFixed(2).replace(".", ",")}</h2>
+                <h3>Deseja alterar o seu saldo?</h3>
                 <input type="number" placeholder="Coloque o valor aqui" value={saldo} onChange={(e) => setSaldo(e.target.value)}></input>
-                <h4>Pronto, agora para terminar, é so clicar em ir</h4>
+                <br/>
+                <br />
                 <button id="butao" type="submit">Ir</button>
                 <button onClick={resetar}>Resetar</button>
                 <button onClick={() => history.push("/history")}>Ir para o histórico</button>
